requirements.txt
python
opencv-python
Pillow
numpy
torch